requirements.txt
pyrogram
tgcrypto
python-decouple
